Refactor sphere.c/.h to hittable.c/.h
