+++ peter_shirley/hittable.h
@@ -9,6 +9,8 @@ typedef struct s_hit_record	t_hit_record;
 typedef struct s_hittable	t_hittable;
 typedef struct s_material	t_material;
 
+typedef t_hittable	t_sphere;
+typedef t_hittable	t_xy_rect;
 typedef t_hittable t_hittable_list;
 typedef t_hittable	t_bvh_node;
 
@@ -46,13 +48,21 @@ struct s_hittable {
 	t_aabb			box;
 };
 
-bool	bounding_box(const t_hittable *self, t_aabb *output_box);
 void	set_face_normal(t_hit_record *self, const t_ray *r, const t_vec3 *outward_normal);
-
-void	hittable_list_add(t_hittable_list *self, t_hittable *object);
 bool	hit(const t_hittable_list *self, const t_ray *r, double t_min, double t_max, t_hit_record *rec);
+bool	bounding_box(const t_hittable *self, t_aabb *output_box);
 
+// new
+t_sphere	sphere_new(t_point cen, double r, t_material *m);
+t_xy_rect	xyrect_new(double x0, double x1, double y0, double y1, double k, t_material *m);
 t_bvh_node	new_bvh_node(t_hittable_list *list);
+
+// alloc
+t_sphere	*sphere_alloc(t_point cen, double r, t_material *m);
+t_xy_rect	*xyrect_alloc(double x0, double x1, double y0, double y1, double k, t_material *m);
 t_bvh_node	*alloc_bvh_node(t_hittable_list *list);
+
+// hittable_list
+void	hittable_list_add(t_hittable_list *self, t_hittable *object);
 
 #endif
